feat: Merge .cpp and .h files
